Workaround for dummy usertab function
This is a workaround for GCC/LLVM optimizing the function call to
usertab which assumes that usertab will not be relocated at runtime.
Providing both a user-defined usertab and the dummy usertab object in
the same executable is possible on Linux, but confuses the C-compilers.

The GCC/autotools projects have been updated, but since VisualStudio
does not include the dummy usertab, the new source-file is not included
for VisualStudio.